Project_Api/requirements.txt
bidict==0.23.1
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
dlib==19.24.4
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
google-auth==2.29.0
google-auth-oauthlib==1.2.0
gspread==6.1.2
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
pillow==10.3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyMySQL==1.1.1
python-engineio==4.9.1
python-socketio==5.11.2
requests==2.32.2
requests-oauthlib==2.0.0
rsa==4.9
simple-websocket==1.0.0
urllib3==2.2.1
Werkzeug==3.0.3
wsproto==1.2.0
